change test case field type (uint64 -> float64)

+++ consensus/ibft/ibft_test.go
@@ -1303,7 +1303,7 @@ func TestQuorumSizeSwitch(t *testing.T) {
 
 	testTable := []struct {
 		name           string
-		switchBlock    uint64
+		switchBlock    float64
 		currentBlock   uint64
 		set            ValidatorSet
 		expectedQuorum int
